refactor(Montant): déplace les conversions toEurosParX dans MontantRécurrent et expose convertitEn
