+++ site/source/domaine/Montant.ts
@@ -47,82 +47,6 @@ export const estEuroParHeure = (
 	montant: Montant
 ): montant is Montant<'€/heure'> => montant.unité === '€/heure'
 
-export const toEurosParMois = (
-	montantRécurrent: Montant<UnitéMonétaireRécurrente>
-): Montant<'€/mois'> => {
-	let valeur = montantRécurrent.valeur
-	switch (montantRécurrent.unité) {
-		case '€/an':
-			valeur = valeur / 12
-			break
-		case '€/jour':
-			valeur = (valeur * 365) / 12
-			break
-		case '€/heure':
-			valeur = (valeur * 24 * 365) / 12
-			break
-	}
-
-	return montant(valeur, '€/mois')
-}
-
-export const toEurosParAn = (
-	montantRécurrent: Montant<UnitéMonétaireRécurrente>
-): Montant<'€/an'> => {
-	let valeur = montantRécurrent.valeur
-	switch (montantRécurrent.unité) {
-		case '€/mois':
-			valeur = valeur * 12
-			break
-		case '€/jour':
-			valeur = valeur * 365
-			break
-		case '€/heure':
-			valeur = valeur * 24 * 365
-			break
-	}
-
-	return montant(valeur, '€/an')
-}
-
-export const toEurosParJour = (
-	montantRécurrent: Montant<UnitéMonétaireRécurrente>
-): Montant<'€/jour'> => {
-	let valeur = montantRécurrent.valeur
-	switch (montantRécurrent.unité) {
-		case '€/an':
-			valeur = valeur / 365
-			break
-		case '€/mois':
-			valeur = (valeur * 12) / 365
-			break
-		case '€/heure':
-			valeur = valeur * 24
-			break
-	}
-
-	return montant(valeur, '€/jour')
-}
-
-export const toEurosParHeure = (
-	montantRécurrent: Montant<UnitéMonétaireRécurrente>
-): Montant<'€/heure'> => {
-	let valeur = montantRécurrent.valeur
-	switch (montantRécurrent.unité) {
-		case '€/an':
-			valeur = valeur / (365 * 24)
-			break
-		case '€/mois':
-			valeur = (valeur * 12) / (365 * 24)
-			break
-		case '€/jour':
-			valeur = valeur / 24
-			break
-	}
-
-	return montant(valeur, '€/heure')
-}
-
 export const montant = <U extends UnitéMonétaire>(
 	valeur: number,
 	unité: U
