+++ site/source/pages/simulateurs/comparaison-statuts/components/ComparaisonListe.tsx
@@ -21,15 +21,14 @@ import {
 	StyledLink,
 	Ul,
 } from '@/design-system'
+import { arrondirÀLEuro, estZéro, montantToString } from '@/domaine/Montant'
 import {
-	arrondirÀLEuro,
-	estZéro,
-	montantToString,
+	plus,
+	pourcentageParRapportÀ,
 	toEurosParAn,
 	toEurosParJour,
 	toEurosParMois,
-} from '@/domaine/Montant'
-import { plus, pourcentageParRapportÀ } from '@/domaine/MontantRécurrent'
+} from '@/domaine/MontantRécurrent'
 import { arrondirÀLUnité, quantitéToString } from '@/domaine/Quantite'
 
 import { CartesStatuts } from './CartesStatuts'
